feat(wc): implement a general cdk overlay for all web components
This will work with any webcomponent by creating an overlay at the root of each
webcomponent

Changed the name of the custom OverlayContainer for clarity
