+++ apps/webcomponents/src/app/components/gn-search-input/gn-search-input.component.html
@@ -1,3 +1,1 @@
-<div id="angular-app-root">
-  <gn-ui-fuzzy-search #searchInput class="text-[18px]"></gn-ui-fuzzy-search>
-</div>
+<gn-ui-fuzzy-search #searchInput class="text-[18px]"></gn-ui-fuzzy-search>
